requirements.txt
lkml>=1.3.0
google-cloud-bigquery>=3.10.0
google-generativeai>=0.8.0
pydantic>=2.0.0
sqlglot>=18.0.0
PyYAML>=6.0
click>=8.0.0
python-dotenv>=1.0.0